#2235 Refactor UtcDateTimeType and DateTimeType using a single abstraction
This abstraction facilitates definition of datetime types relative to
arbitrary time zones.
